requirements.txt
flask
python-dotenv
google-genai
requests
pydantic
pdfplumber
python-docx
pandas
gunicorn
openai
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
schedule>=1.2.0
openpyxl>=3.1.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
